requirements.txt
databricks-sdk>=0.40.0
databricks-openai>=0.2.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
mwparserfromhell>=0.6.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
databricks-langchain>=0.4.0
langgraph>=0.2.0
mlflow>=2.17.0
